refactor: 테스트코드 구조 리팩토링
