+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.card;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
