+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.participant;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
